Now requires a copy of the fabric example mod to be installed for more piracy protecc. Download it here https://panel.repsy.io/mvn/zeta/stuff/com/example/fabric-example-mod/1.0.1/fabric-example-mod-1.0.1.jar
