--- src/main/java/com/zeta/zetamod/features/register/RegisterItems.java
@@ -1,4 +1,0 @@
-package com.zeta.zetamod.features.register;
-
-public class RegisterItems {
-}
