+++ src/main/java/zetamod/mod/features/biomes/ObsilandBiome.java
@@ -1,15 +1,9 @@
-package com.zeta.zetamod.features.biomes;
+package zetamod.mod.features.biomes;
 
-import com.zeta.zetamod.main.ZMMain;
+import zetamod.mod.main.ZetaModMain;
 import net.fabricmc.fabric.api.biome.v1.OverworldBiomes;
 import net.fabricmc.fabric.api.biome.v1.OverworldClimate;
-import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
-import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
-import net.minecraft.block.Material;
-import net.minecraft.item.BlockItem;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemGroup;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.BuiltinRegistries;
 import net.minecraft.util.registry.Registry;
@@ -24,19 +18,16 @@ import net.minecraft.world.gen.surfacebuilder.SurfaceBuilder;
 import net.minecraft.world.gen.surfacebuilder.TernarySurfaceConfig;
 import org.apache.logging.log4j.Level;
 
-public class ConcerningBiome {
-
-    public static final Block GIGACONCERN_BLOCK = new Block(FabricBlockSettings.of(Material.METAL).hardness(8.0f));
-
-    private static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> CONCERNING_SURFACE_BUILDER = SurfaceBuilder.DEFAULT
+public class ObsilandBiome {
+    private static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> OBSIDIAN_SURFACE_BUILDER = SurfaceBuilder.DEFAULT
             .withConfig(new TernarySurfaceConfig(
-                    ZMMain.CONCERN_BLOCK.getDefaultState(),
-                    ZMMain.HYPERCONCERN_BLOCK.getDefaultState(),
-                    GIGACONCERN_BLOCK.getDefaultState()));
+                    Blocks.OBSIDIAN.getDefaultState(),
+                    Blocks.DIRT.getDefaultState(),
+                    Blocks.GRAVEL.getDefaultState()));
 
-    private static final Biome CONCERNINGS = createConcernings();
+    private static final Biome OBSILAND = createObsiland();
 
-    private static Biome createConcernings() {
+    private static Biome createObsiland() {
         // We specify what entities spawn and what features generate in the biome.
         // Aside from some structures, trees, rocks, plants and
         //   custom entities, these are mostly the same for each biome.
@@ -47,7 +38,7 @@ public class ConcerningBiome {
         DefaultBiomeFeatures.addMonsters(spawnSettings, 95, 5, 100);
 
         GenerationSettings.Builder generationSettings = new GenerationSettings.Builder();
-        generationSettings.surfaceBuilder(CONCERNING_SURFACE_BUILDER);
+        generationSettings.surfaceBuilder(OBSIDIAN_SURFACE_BUILDER);
         DefaultBiomeFeatures.addDefaultUndergroundStructures(generationSettings);
         DefaultBiomeFeatures.addLandCarvers(generationSettings);
         DefaultBiomeFeatures.addDefaultLakes(generationSettings);
@@ -76,28 +67,23 @@ public class ConcerningBiome {
                 .build();
     }
 
-    private static final String MOD_ID = ZMMain.MOD_ID;
-    public static final RegistryKey<Biome> CONCERNING_KEY = RegistryKey.of(Registry.BIOME_KEY, new Identifier(MOD_ID, "concerning"));
+    private static final String MOD_ID = ZetaModMain.MOD_ID;
+    public static final RegistryKey<Biome> OBSILAND_KEY = RegistryKey.of(Registry.BIOME_KEY, new Identifier(MOD_ID, "obsiland"));
 
     public static void register() {
-        ZMMain.log(Level.INFO, "Adding biomes");
-        Registry.register(BuiltinRegistries.CONFIGURED_SURFACE_BUILDER, new Identifier(MOD_ID, "concerning"), CONCERNING_SURFACE_BUILDER);
-        Registry.register(BuiltinRegistries.BIOME, CONCERNING_KEY.getValue(), CONCERNINGS);
-        boolean addToWorldgenBoolean = true;
+        ZetaModMain.log(Level.INFO, "Adding biomes");
+        Registry.register(BuiltinRegistries.CONFIGURED_SURFACE_BUILDER, new Identifier(MOD_ID, "obsidian"), OBSIDIAN_SURFACE_BUILDER);
+        Registry.register(BuiltinRegistries.BIOME, OBSILAND_KEY.getValue(), OBSILAND);
+        boolean addToWorldgenBoolean = false;
         if (addToWorldgenBoolean) {
             addToWorldgen();
         }
-        ZMMain.log(Level.INFO,
-                "Initialized easter egg biome, hehe."
-                        //"Thanks to https://misode.github.io/ for the custom dimension creator I used to make the Lush Nether."
+        ZetaModMain.log(Level.INFO,
+                "Initialized Obsiland biome. Thanks to https://misode.github.io/ for the custom dimension creator I used to make the Lush Nether."
         );
-        Registry.register(Registry.BLOCK, new Identifier(MOD_ID, "gigaconcern_block"), GIGACONCERN_BLOCK);
-        Registry.register(Registry.ITEM, new Identifier(MOD_ID, "gigaconcern_block"), new BlockItem(GIGACONCERN_BLOCK, new Item.Settings().group(ItemGroup.MISC)));
-
     }
-    public static double conc = ZMMain.concerning_weight;
     public static void addToWorldgen() {
-        OverworldBiomes.addContinentalBiome(CONCERNING_KEY, OverworldClimate.TEMPERATE, conc);
-        OverworldBiomes.addContinentalBiome(CONCERNING_KEY, OverworldClimate.COOL, conc);
+        OverworldBiomes.addContinentalBiome(OBSILAND_KEY, OverworldClimate.TEMPERATE, 2D);
+        OverworldBiomes.addContinentalBiome(OBSILAND_KEY, OverworldClimate.COOL, 2D);
     }
 }
