+++ src/main/java/zetamod/mod/features/items/CrashItemInitializer.java
@@ -1,6 +1,6 @@
-package com.zeta.zetamod.features.items;
+package zetamod.mod.features.items;
 
-import com.zeta.zetamod.main.ZMMain;
+import zetamod.mod.main.ZetaModMain;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.minecraft.item.ItemGroup;
@@ -12,7 +12,7 @@ public class CrashItemInitializer implements ModInitializer {
 
     @Override
     public void onInitialize() {
-        Registry.register(Registry.ITEM, new Identifier(ZMMain.MOD_ID, "crashitem"), new CrashItem(new FabricItemSettings().group(ItemGroup.MISC)));
-        ZMMain.log(Level.INFO,"Intialized debug item!");
+        Registry.register(Registry.ITEM, new Identifier(ZetaModMain.MOD_ID, "crashitem"), new CrashItem(new FabricItemSettings().group(ItemGroup.MISC)));
+        ZetaModMain.log(Level.INFO,"Intialized debug item!");
     }
 }
