+++ src/main/java/zetamod/mod/features/register/RegisterItems.java
@@ -1,0 +1,4 @@
+package zetamod.mod.features.register;
+
+public class RegisterItems {
+}
