+++ src/main/java/zetamod/mod/mixins/worldborder/MixinAquiferSampler.java
@@ -1,4 +1,4 @@
-package com.zeta.zetamod.mixins.worldborder;
+package zetamod.mod.mixins.worldborder;
 
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
