+++ src/main/java/zetamod/mod/mixins/worldborder/MixinEntity.java
@@ -1,15 +1,15 @@
-package com.zeta.zetamod.mixins.worldborder;
+package zetamod.mod.mixins.worldborder;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
-import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.entity.Entity;
 import net.minecraft.util.math.MathHelper;
 
-@Mixin(PlayerEntity.class)
-public class MixinPlayerEntity {
-	@Redirect(method = "tick", at = @At(target = "Lnet/minecraft/util/math/MathHelper;clamp(DDD)D", value = "INVOKE"))
+@Mixin(Entity.class)
+public class MixinEntity {
+	@Redirect(method = "updatePosition", at = @At(target = "Lnet/minecraft/util/math/MathHelper;clamp(DDD)D", value = "INVOKE"))
 	private double redirectClamp(double value, double min, double max) {
 		return MathHelper.clamp(value, Integer.MIN_VALUE, Integer.MAX_VALUE);
 	}
