+++ src/main/resources/zetamod.mixins.json
@@ -4,16 +4,16 @@
   "package": "com.zeta.zetamod.mixins",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "FarLandsMixins",
-    "worldborder.MixinEntity",
-    "worldborder.MixinMinecraftServer",
-    "worldborder.MixinNoiseSamplingConfig",
-    "worldborder.MixinPlayerEntity",
-    "worldborder.MixinServerPlayNetworkHandler",
-    "worldborder.MixinWorld",
-    "worldborder.MixinWorldBorder",
-    "worldborder.MixinWorldBorderProperties",
-    "worldborder.MixinAquiferSampler"
+    "zetamod.mod.mixins.FarLandsMixins",
+    "zetamod.mod.mixins.worldborder.MixinEntity",
+    "zetamod.mod.mixins.worldborder.MixinMinecraftServer",
+    "zetamod.mod.mixins.worldborder.MixinNoiseSamplingConfig",
+    "zetamod.mod.mixins.worldborder.MixinPlayerEntity",
+    "zetamod.mod.mixins.worldborder.MixinServerPlayNetworkHandler",
+    "zetamod.mod.mixins.worldborder.MixinWorld",
+    "zetamod.mod.mixins.worldborder.MixinWorldBorder",
+    "zetamod.mod.mixins.worldborder.MixinWorldBorderProperties",
+    "zetamod.mod.mixins.worldborder.MixinAquiferSampler"
   ],
   "client": [
